Refactor dir and vdir code
